PICARD-1671: Moved tags compatibility options to separate page
