+++ picard/ui/ui_options_tags.py
@@ -10,7 +10,7 @@ from PyQt5 import QtCore, QtGui, QtWidgets
 class Ui_TagsOptionsPage(object):
     def setupUi(self, TagsOptionsPage):
         TagsOptionsPage.setObjectName("TagsOptionsPage")
-        TagsOptionsPage.resize(539, 633)
+        TagsOptionsPage.resize(539, 525)
         self.vboxlayout = QtWidgets.QVBoxLayout(TagsOptionsPage)
         self.vboxlayout.setObjectName("vboxlayout")
         self.write_tags = QtWidgets.QCheckBox(TagsOptionsPage)
@@ -46,122 +46,8 @@ class Ui_TagsOptionsPage(object):
         self.preserved_tags_help.setObjectName("preserved_tags_help")
         self.vboxlayout1.addWidget(self.preserved_tags_help)
         self.vboxlayout.addWidget(self.before_tagging)
-        self.tag_compatibility = QtWidgets.QGroupBox(TagsOptionsPage)
-        self.tag_compatibility.setObjectName("tag_compatibility")
-        self.vboxlayout2 = QtWidgets.QVBoxLayout(self.tag_compatibility)
-        self.vboxlayout2.setContentsMargins(-1, 6, -1, 7)
-        self.vboxlayout2.setSpacing(2)
-        self.vboxlayout2.setObjectName("vboxlayout2")
-        self.id3v2_version = QtWidgets.QGroupBox(self.tag_compatibility)
-        self.id3v2_version.setFlat(False)
-        self.id3v2_version.setCheckable(False)
-        self.id3v2_version.setObjectName("id3v2_version")
-        self.horizontalLayout = QtWidgets.QHBoxLayout(self.id3v2_version)
-        self.horizontalLayout.setContentsMargins(-1, 6, -1, 7)
-        self.horizontalLayout.setObjectName("horizontalLayout")
-        self.write_id3v24 = QtWidgets.QRadioButton(self.id3v2_version)
-        self.write_id3v24.setChecked(True)
-        self.write_id3v24.setObjectName("write_id3v24")
-        self.horizontalLayout.addWidget(self.write_id3v24)
-        self.write_id3v23 = QtWidgets.QRadioButton(self.id3v2_version)
-        self.write_id3v23.setChecked(False)
-        self.write_id3v23.setObjectName("write_id3v23")
-        self.horizontalLayout.addWidget(self.write_id3v23)
-        self.label = QtWidgets.QLabel(self.id3v2_version)
-        self.label.setText("")
-        self.label.setWordWrap(True)
-        self.label.setObjectName("label")
-        self.horizontalLayout.addWidget(self.label)
-        spacerItem1 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem1)
-        self.vboxlayout2.addWidget(self.id3v2_version)
-        self.id3v2_text_encoding = QtWidgets.QGroupBox(self.tag_compatibility)
-        self.id3v2_text_encoding.setObjectName("id3v2_text_encoding")
-        self.horizontalLayout_2 = QtWidgets.QHBoxLayout(self.id3v2_text_encoding)
-        self.horizontalLayout_2.setContentsMargins(-1, 6, -1, 7)
-        self.horizontalLayout_2.setObjectName("horizontalLayout_2")
-        self.enc_utf8 = QtWidgets.QRadioButton(self.id3v2_text_encoding)
-        self.enc_utf8.setObjectName("enc_utf8")
-        self.horizontalLayout_2.addWidget(self.enc_utf8)
-        self.enc_utf16 = QtWidgets.QRadioButton(self.id3v2_text_encoding)
-        self.enc_utf16.setObjectName("enc_utf16")
-        self.horizontalLayout_2.addWidget(self.enc_utf16)
-        self.enc_iso88591 = QtWidgets.QRadioButton(self.id3v2_text_encoding)
-        self.enc_iso88591.setObjectName("enc_iso88591")
-        self.horizontalLayout_2.addWidget(self.enc_iso88591)
-        spacerItem2 = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout_2.addItem(spacerItem2)
-        self.label_2 = QtWidgets.QLabel(self.id3v2_text_encoding)
-        self.label_2.setText("")
-        self.label_2.setWordWrap(True)
-        self.label_2.setObjectName("label_2")
-        self.horizontalLayout_2.addWidget(self.label_2)
-        self.vboxlayout2.addWidget(self.id3v2_text_encoding)
-        self.hbox_id3v23_join_with = QtWidgets.QHBoxLayout()
-        self.hbox_id3v23_join_with.setObjectName("hbox_id3v23_join_with")
-        self.label_id3v23_join_with = QtWidgets.QLabel(self.tag_compatibility)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
-        sizePolicy.setHorizontalStretch(4)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label_id3v23_join_with.sizePolicy().hasHeightForWidth())
-        self.label_id3v23_join_with.setSizePolicy(sizePolicy)
-        self.label_id3v23_join_with.setObjectName("label_id3v23_join_with")
-        self.hbox_id3v23_join_with.addWidget(self.label_id3v23_join_with)
-        self.id3v23_join_with = QtWidgets.QComboBox(self.tag_compatibility)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(1)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.id3v23_join_with.sizePolicy().hasHeightForWidth())
-        self.id3v23_join_with.setSizePolicy(sizePolicy)
-        self.id3v23_join_with.setEditable(True)
-        self.id3v23_join_with.setObjectName("id3v23_join_with")
-        self.id3v23_join_with.addItem("")
-        self.id3v23_join_with.setItemText(0, "/")
-        self.id3v23_join_with.addItem("")
-        self.id3v23_join_with.setItemText(1, "; ")
-        self.id3v23_join_with.addItem("")
-        self.id3v23_join_with.setItemText(2, " / ")
-        self.hbox_id3v23_join_with.addWidget(self.id3v23_join_with)
-        self.vboxlayout2.addLayout(self.hbox_id3v23_join_with)
-        self.itunes_compatible_grouping = QtWidgets.QCheckBox(self.tag_compatibility)
-        self.itunes_compatible_grouping.setObjectName("itunes_compatible_grouping")
-        self.vboxlayout2.addWidget(self.itunes_compatible_grouping)
-        self.write_id3v1 = QtWidgets.QCheckBox(self.tag_compatibility)
-        self.write_id3v1.setObjectName("write_id3v1")
-        self.vboxlayout2.addWidget(self.write_id3v1)
-        self.vboxlayout.addWidget(self.tag_compatibility)
-        self.aac_tags = QtWidgets.QGroupBox(TagsOptionsPage)
-        self.aac_tags.setObjectName("aac_tags")
-        self.verticalLayout = QtWidgets.QVBoxLayout(self.aac_tags)
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.aac_save_ape = QtWidgets.QRadioButton(self.aac_tags)
-        self.aac_save_ape.setChecked(True)
-        self.aac_save_ape.setObjectName("aac_save_ape")
-        self.verticalLayout.addWidget(self.aac_save_ape)
-        self.aac_no_tags = QtWidgets.QRadioButton(self.aac_tags)
-        self.aac_no_tags.setObjectName("aac_no_tags")
-        self.verticalLayout.addWidget(self.aac_no_tags)
-        self.remove_ape_from_aac = QtWidgets.QCheckBox(self.aac_tags)
-        self.remove_ape_from_aac.setObjectName("remove_ape_from_aac")
-        self.verticalLayout.addWidget(self.remove_ape_from_aac)
-        self.vboxlayout.addWidget(self.aac_tags)
-        self.ac3_files = QtWidgets.QGroupBox(TagsOptionsPage)
-        self.ac3_files.setObjectName("ac3_files")
-        self.verticalLayout_2 = QtWidgets.QVBoxLayout(self.ac3_files)
-        self.verticalLayout_2.setObjectName("verticalLayout_2")
-        self.ac3_save_ape = QtWidgets.QRadioButton(self.ac3_files)
-        self.ac3_save_ape.setChecked(True)
-        self.ac3_save_ape.setObjectName("ac3_save_ape")
-        self.verticalLayout_2.addWidget(self.ac3_save_ape)
-        self.ac3_no_tags = QtWidgets.QRadioButton(self.ac3_files)
-        self.ac3_no_tags.setObjectName("ac3_no_tags")
-        self.verticalLayout_2.addWidget(self.ac3_no_tags)
-        self.remove_ape_from_ac3 = QtWidgets.QCheckBox(self.ac3_files)
-        self.remove_ape_from_ac3.setObjectName("remove_ape_from_ac3")
-        self.verticalLayout_2.addWidget(self.remove_ape_from_ac3)
-        self.vboxlayout.addWidget(self.ac3_files)
-        spacerItem3 = QtWidgets.QSpacerItem(274, 41, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
-        self.vboxlayout.addItem(spacerItem3)
+        spacerItem1 = QtWidgets.QSpacerItem(274, 41, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
+        self.vboxlayout.addItem(spacerItem1)
 
         self.retranslateUi(TagsOptionsPage)
         QtCore.QMetaObject.connectSlotsByName(TagsOptionsPage)
@@ -170,13 +56,6 @@ class Ui_TagsOptionsPage(object):
         TagsOptionsPage.setTabOrder(self.clear_existing_tags, self.remove_id3_from_flac)
         TagsOptionsPage.setTabOrder(self.remove_id3_from_flac, self.remove_ape_from_mp3)
         TagsOptionsPage.setTabOrder(self.remove_ape_from_mp3, self.preserved_tags)
-        TagsOptionsPage.setTabOrder(self.preserved_tags, self.write_id3v24)
-        TagsOptionsPage.setTabOrder(self.write_id3v24, self.write_id3v23)
-        TagsOptionsPage.setTabOrder(self.write_id3v23, self.enc_utf8)
-        TagsOptionsPage.setTabOrder(self.enc_utf8, self.enc_utf16)
-        TagsOptionsPage.setTabOrder(self.enc_utf16, self.enc_iso88591)
-        TagsOptionsPage.setTabOrder(self.enc_iso88591, self.id3v23_join_with)
-        TagsOptionsPage.setTabOrder(self.id3v23_join_with, self.write_id3v1)
 
     def retranslateUi(self, TagsOptionsPage):
         _translate = QtCore.QCoreApplication.translate
@@ -188,23 +67,3 @@ class Ui_TagsOptionsPage(object):
         self.remove_ape_from_mp3.setText(_("Remove APEv2 tags from MP3 files"))
         self.preserved_tags_label.setText(_("Preserve these tags from being cleared or overwritten with MusicBrainz data:"))
         self.preserved_tags_help.setText(_("Tags are separated by commas, and are case-sensitive."))
-        self.tag_compatibility.setTitle(_("Tag Compatibility"))
-        self.id3v2_version.setTitle(_("ID3v2 Version"))
-        self.write_id3v24.setText(_("2.4"))
-        self.write_id3v23.setText(_("2.3"))
-        self.id3v2_text_encoding.setTitle(_("ID3v2 Text Encoding"))
-        self.enc_utf8.setText(_("UTF-8"))
-        self.enc_utf16.setText(_("UTF-16"))
-        self.enc_iso88591.setText(_("ISO-8859-1"))
-        self.label_id3v23_join_with.setText(_("Join multiple ID3v2.3 tags with:"))
-        self.id3v23_join_with.setToolTip(_("<html><head/><body><p>Default is \'/\' to maintain compatibility with previous Picard releases.</p><p>New alternatives are \';_\' or \'_/_\' or type your own. </p></body></html>"))
-        self.itunes_compatible_grouping.setText(_("Save iTunes compatible grouping and work"))
-        self.write_id3v1.setText(_("Also include ID3v1 tags in the files"))
-        self.aac_tags.setTitle(_("AAC files"))
-        self.aac_save_ape.setText(_("Save APEv2 tags"))
-        self.aac_no_tags.setText(_("Do not save tags"))
-        self.remove_ape_from_aac.setText(_("Remove APEv2 tags from AAC files"))
-        self.ac3_files.setTitle(_("AC3 files"))
-        self.ac3_save_ape.setText(_("Save APEv2 tags"))
-        self.ac3_no_tags.setText(_("Do not save tags"))
-        self.remove_ape_from_ac3.setText(_("Remove APEv2 tags from AC3 files"))
